update dragonTraining module

+++ server/src/modules/dragon/dragonActions.ts
@@ -18,7 +18,7 @@ const browse: RequestHandler = async (req, res, next) => {
 };
 
 const browseByUser: RequestHandler = async (req, res, next) => {
-  const userId = Number(req.params.id);
+  const userId = Number(req.params.userId);
 
   try {
     const dragons = await dragonRepository.readAllByUser(userId);
@@ -35,10 +35,11 @@ const browseByUser: RequestHandler = async (req, res, next) => {
 
 // R of BREAD
 const read: RequestHandler = async (req, res, next) => {
-  const dragonId = Number(req.params.id);
+  const userId = Number(req.params.userId);
+  const dragonId = Number(req.params.dragonId);
 
   try {
-    const dragon = await dragonRepository.read(dragonId);
+    const dragon = await dragonRepository.read({ userId, dragonId });
 
     if (!dragon) {
       res.sendStatus(404);
@@ -52,14 +53,16 @@ const read: RequestHandler = async (req, res, next) => {
 
 // E of BREAD
 const edit: RequestHandler = async (req, res, next) => {
-  const dragonId = Number(req.params.id);
+  const userId = Number(req.params.userId);
+  const dragonId = Number(req.params.dragonId);
 
   try {
-    const dragon: Omit<Dragon, "specie_id" | "user_id"> = {
+    const dragon: Omit<Dragon, "specie_id"> = {
       name: req.body.name,
       strength: req.body.strength,
       speed: req.body.speed,
       stamina: req.body.stamina,
+      user_id: userId,
       id: dragonId,
     };
 
@@ -77,11 +80,13 @@ const edit: RequestHandler = async (req, res, next) => {
 
 // A of BREAD
 const add: RequestHandler = async (req, res, next) => {
+  const userId = Number(req.params.userId);
+
   try {
     const dragon: Omit<Dragon, "id" | "strength" | "speed" | "stamina"> = {
       name: req.body.name,
       specie_id: req.body.specie_id,
-      user_id: req.body.user_id,
+      user_id: userId,
     };
 
     const insertId = await dragonRepository.create(dragon);
@@ -94,10 +99,11 @@ const add: RequestHandler = async (req, res, next) => {
 
 // D of BREAD
 const destroy: RequestHandler = async (req, res, next) => {
-  const dragonId = Number(req.params.id);
+  const userId = Number(req.params.userId);
+  const dragonId = Number(req.params.dragonId);
 
   try {
-    const affectedRows = await dragonRepository.destroy(dragonId);
+    const affectedRows = await dragonRepository.destroy({ userId, dragonId });
 
     if (affectedRows === 0) {
       res.sendStatus(404);
